+++ Scripts/PatchCopyUnmoddedSaveFilesIfNeeded.cs
@@ -17,7 +17,7 @@ public static class PatchCopyUnmoddedSaveFilesIfNeeded
     public static bool Prefix()
     {
         // 如果当前不是 gameplay modded 状态, 则直接跳过原始方法, 避免首次存档复制
-        if (!RespectAffectsGameplayMod.IsEffectivelyModded())
+        if (!RespectAffectsGameplayMod.IsEffectivelyModded(true))
         {
             ModLog.Info("非 gameplay modded 状态, 跳过首次存档复制");
             return false;